refactor(files): Reorganise Svelte components.
