+++ src/lib/components/game/Achievements.svelte
