+++ src/lib/components/game/Atom.svelte
