+++ src/lib/components/game/Canvas.svelte
