+++ src/lib/components/game/SkillNode.svelte
