+++ src/lib/components/prestige/BonusPhoton.svelte
