+++ src/lib/components/prestige/PhotonCounter.svelte
