+++ src/lib/components/ui/AutoButton.svelte
