+++ src/lib/components/ui/Currency.svelte
