+++ src/lib/components/ui/NotificationDot.svelte
