+++ src/lib/components/ui/StatItem.svelte
